Add mixed auth example
