--- ecommerce_server_python/sample_data.json
@@ -1,9 +1,0 @@
-{
-  "products": [
-    {
-      "id": "matcha-pizza",
-      "image": "https://persistent.oaistatic.com/pizzaz-cart-xl/matcha-pizza.png",
-      "quantity": 1
-    }
-  ]
-}
